pyproject.toml
[tool.black]
line-length=79

[tool.poetry]
name = "starboard"
version = "0.0.0"
description = ""
authors = ["Circuit"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
hikari = "^2.0.0-alpha.105"
hikari-tanjun = "^2.3.0-alpha.1"
hikari-clusters = {git = "https://github.com/TrigonDev/hikari-clusters.git", rev = "main"}
apgorm = {git = "https://github.com/TrigonDev/apgorm.git", rev = "main"}
aiohttp = "^3.8.1"
cachetools = "^5.0.0"
types-cachetools = "^4.2.9"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
isort = "^5.9.3"
mypy = "^0.931"
flake8 = "^4.0.1"
pre-commit = "^2.15.0"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
